feat: add blocks search endpoint
