+++ .gitignore
@@ -38,3 +38,4 @@ dist
 site/public/blocks
 site/site-map.json
 site/yarn.lock
+site/blocks-data.json
