+++ site/src/pages/index.page.tsx
@@ -6,7 +6,7 @@ import { Section2 } from "../components/pages/home/Section2";
 import { WhyBlockProtocol1Section } from "../components/pages/home/WhyBlockProtocol1Section";
 import { WhyBlockProtocol2Section } from "../components/pages/home/WhyBlockProtocol2Section";
 import { RegistrySection } from "../components/pages/home/RegistrySection";
-import { BlockMetadata, readBlocksFromDisk } from "./api/blocks.api";
+import { BlockMetadata, readBlocksFromDisk } from "../lib/blocks";
 
 export const HOME_PAGE_HEADER_HEIGHT = 750;
